refactor: :recycle: refactor commission value
Refactoring the commission value from u64 to u128 to match the coin value Uint128.

+++ src/msg.rs
@@ -9,7 +9,7 @@ pub struct InstantiateMsg {
     pub denom: String,
     pub owner: Option<String>,
     pub description: String,
-    pub commission: Option<u64>,
+    pub commission: Option<u128>,
 }
 
 #[cw_serde]
