+++ src/state.rs
@@ -19,7 +19,7 @@ pub struct Config {
     // auction description
     pub description: String,
     // commission on each valid bid
-    pub commission: u64,
+    pub commission: u128,
 }
 
 // Auction Current status
